requirements.txt
asgiref
crispy-bootstrap5
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
environ
Markdown
pillow
psycopg2
psycopg2-binary
sqlparse
typing_extensions
tzdata
